refactor(auth): 🔥 remove deprecated user authentication functions and update imports
Replace getUserAuthStatus with checkUserStatusClient in components
