+++ src/components/Header.astro
@@ -1,7 +1,7 @@
 ---
-import { getUserAuthStatus } from "@/auth/user.service";
 import { type Lang, useTranslatedPath, useTranslations } from "@/i18n";
 import { cn } from "@/lib/utils";
+import { checkUserStatusClient } from "@/models/auth/user.service";
 import {
 	type AuthInfo,
 	dropdownMenuItems,
@@ -25,7 +25,7 @@ const translatePath = useTranslatedPath(lang);
 interface Props extends HTMLAttributes<"header"> {}
 const { ...attrs } = Astro.props as Props;
 
-const isLoggedIn = await getUserAuthStatus(Astro.request);
+const isLoggedIn = await checkUserStatusClient();
 
 const filteredMainMenuItems = filterMenuItems(mainMenuItems);
 const filteredDropdownMenuItems = filterMenuItems(dropdownMenuItems);
@@ -112,7 +112,7 @@ const authInfo: AuthInfo = {
   </style>
 
   <script>
-    import { checkUserStatusClient } from "../auth/user.service";
+    import { checkUserStatusClient } from "@/models/auth/user.service";
 
     // Handle colored-hero classes for button styling
     const header = document.getElementById('header');
